FastAPI Basic Setup/requirements.txt
fastapi[all]
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
marshmallow
pymysql
pytest